# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-commerce platform."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "example": [
                  { "id": "64f1c72d8a1f3c4567d89ab1", "name": "John Doe", "email": "john@example.com", "role": "user" },
                  { "id": "64f1c72d8a1f3c4567d89ab2", "name": "Jane Admin", "email": "jane@example.com", "role": "admin" }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" },
              "example": { "name": "Alice Johnson", "email": "alice@example.com", "password": "Password123!", "role": "user" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "example": { "id": "64f1c72d8a1f3c4567d89ab9", "name": "Alice Johnson", "email": "alice@example.com", "role": "user" }
              }
            }
          },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/users/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "summary": "Get user by id",
        "tags": ["Users"],
        "responses": {
          "200": { "description": "User found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" }, "example": { "id": "64f1c72d8a1f3c4567d89ab1", "name": "John Doe", "email": "john@example.com", "role": "user" } } } },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update a user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" }, "example": { "name": "John Updated", "address": "22 New Ave" } } }
        },
        "responses": {
          "200": { "description": "User updated", "content": { "application/json": { "example": { "id": "64f1c72d8a1f3c4567d89ab1", "name": "John Updated", "email": "john@example.com" } } } },
          "400": { "description": "Invalid request" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "tags": ["Users"],
        "responses": {
          "200": { "description": "User deleted", "content": { "application/json": { "example": { "message": "User deleted" } } } },
          "404": { "description": "User not found" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "6500b12f9d12345678c9ab30",
                    "name": "iPhone 14 Pro",
                    "description": "Latest Apple smartphone",
                    "price": 1200,
                    "stock": 15,
                    "category": "Electronics",
                    "images": ["https://example.com/iphone14.jpg"],
                    "isFeatured": true
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product (admin)",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": { "schema": { "$ref": "#/components/schemas/ProductCreateMultipart" } },
            "application/json": { "schema": { "$ref": "#/components/schemas/ProductCreate" }, "example": { "name": "Samsung Galaxy S23", "description": "High-end Android smartphone", "price": 1000, "stock": 25, "category": "Electronics", "images": ["https://example.com/galaxys23.jpg"], "isFeatured": false } }
          }
        },
        "responses": {
          "201": { "description": "Product created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" }, "example": { "id": "6500b22f9d12345678c9ab33", "name": "Samsung Galaxy S23", "price": 1000 } } } },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/products/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "summary": "Get product by id",
        "tags": ["Products"],
        "responses": {
          "200": { "description": "Product found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "summary": "Update a product",
        "tags": ["Products"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductUpdate" }, "example": { "price": 950, "stock": 30 } } }
        },
        "responses": { "200": { "description": "Product updated", "content": { "application/json": { "example": { "id": "6500b22f9d12345678c9ab33", "name": "Samsung Galaxy S23", "price": 950, "stock": 30 } } } }, "404": { "description": "Product not found" } }
      },
      "delete": {
        "summary": "Delete a product",
        "tags": ["Products"],
        "responses": { "200": { "description": "Product deleted", "content": { "application/json": { "example": { "message": "Product deleted" } } } }, "404": { "description": "Product not found" } }
      }
    },
...
    "/payments": {
      "get": {
        "summary": "Get all payments",
        "tags": ["Payments"],
        "responses": { "200": { "description": "List of payments", "content": { "application/json": { "example": [ { "id": "6500e52f9d12345678c9ab66", "order": "6500c32f9d12345678c9ab44", "status": "pending", "amount": 2000 } ] } } } }
      },
      "post": {
        "summary": "Create a new payment (record)",
        "tags": ["Payments"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentCreate" }, "example": { "order": "6500c32f9d12345678c9ab44", "paymentId": "PAY-12345", "status": "completed", "amount": 2000, "paymentMethod": "paypal" } } } },
        "responses": { "201": { "description": "Payment created", "content": { "application/json": { "example": { "id": "6500e52f9d12345678c9ab66", "order": "6500c32f9d12345678c9ab44", "status": "completed", "amount": 2000 } } } } }
      }
    },
    "/payments/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": { "summary": "Get payment by id", "tags": ["Payments"], "responses": { "200": { "description": "Payment found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Payment" } } } }, "404": { "description": "Payment not found" } } },
      "put": { "summary": "Update payment", "tags": ["Payments"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentUpdate" }, "example": { "status": "refunded" } } } }, "responses": { "200": { "description": "Payment updated" }, "404": { "description": "Payment not found" } } },
      "delete": { "summary": "Delete payment", "tags": ["Payments"], "responses": { "200": { "description": "Payment deleted" }, "404": { "description": "Payment not found" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "UserCreate": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin", "subadmin"], "default": "user" },
          "isOnline": { "type": "boolean" },
          "lastSeen": { "type": "string", "format": "date-time" },
          "address": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "email": { "type": "string" }, "role": { "type": "string" } }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string" } },
          "isFeatured": { "type": "boolean" }
        }
      },
      "ProductCreate": {
        "type": "object",
        "required": ["name", "price", "stock"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "isFeatured": { "type": "boolean" }
        }
      },
      "ProductCreateMultipart": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string" },
          "image": { "type": "string", "format": "binary" }
        },
        "required": ["name", "price"]
      },
      "ProductUpdate": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "price": { "type": "number" }, "stock": { "type": "integer" }, "category": { "type": "string" }, "images": { "type": "array", "items": { "type": "string" } } }
      },
...
      "PaymentCreate": {
        "type": "object",
        "required": ["order", "amount", "paymentMethod"],
        "properties": {
          "order": { "type": "string" },
          "paymentId": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "completed", "failed"] },
          "amount": { "type": "number" },
          "paymentMethod": { "type": "string" }
        }
      },
      "Payment": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "order": { "type": "string" }, "paymentId": { "type": "string" }, "status": { "type": "string" }, "amount": { "type": "number" }, "paymentMethod": { "type": "string" } }
      },
      "PaymentUpdate": { "type": "object", "properties": { "status": { "type": "string" } } },
...
    }
  }
}